Increase size of representation of variables

+++ src/001/001.cpp
@@ -3,14 +3,14 @@
 #define LIMIT 1000
 
 int main() {
-    unsigned result = 5;
-    for (unsigned i = 3; i < LIMIT; i += 6) {
-        unsigned j = i;
+    unsigned long long result = 5;
+    for (unsigned long long i = 3; i < LIMIT; i += 6) {
+        unsigned long long j = i;
         while (j < LIMIT) {
             result += j;
             if (!(j % 5)) {
                 result += (j-5);
-                if (j+5 < LIMIT) {
+                if (j + 5 < LIMIT) {
                     result += (j+5);
                 }
             }
@@ -18,6 +18,6 @@ int main() {
         }
     }
 
-    printf("Result is %u\n", result);
+    printf("Result is %llu\n", result);
     return 0;
 }
